requirements.txt
streamlit>=1.30
pandas>=2.0
numpy>=1.24
scikit-learn>=1.4
plotly>=5.18
scipy>=1.10
requests>=2.31
python-dotenv>=1.0
